Throttle main-threaded animation on OS X.
Summary:
dispatch_async has no throttling mechanism. If your main thread is contended, the high-priority background thread will keep enqueueing blocks on the main thread and you'll end up with a full queue of blocks and the app will start getting behind with little possibility of recovery.o

Test plan:
Run a mac app whose main thread is contended, add an else case and a log when we don't enqueue a dispatch. Observe that it is called in those cases.