Allows animations to be paused *now*
Calls to _render will automatically enable pause, this wouldn't be an issue but the call to _render is called at some time in the future (especially if CADisplayLink is not configured).  This patch allows you to use state.paused = true at creation time.  By default, the animation will start in a non-paused state